+++ lib/generators/templates/helpers/amp/css_tag_helper.rb
@@ -1,0 +1,7 @@
+module ApplicationHelper
+
+    def amp_css
+        content_for :css
+    end
+
+end